# Performance optimization: Significant improvements for chaining and batch operations (v4.2.0)