--- a/Rainbow.xcodeproj/project.pbxproj
+++ b/Rainbow.xcodeproj/project.pbxproj
@@ -45,6 +45,12 @@
 		4BC821811C29254C00F68E70 /* Rainbow.framework in Frameworks */ = {isa = PBXBuildFile; fileRef = 4BC821761C29254C00F68E70 /* Rainbow.framework */; };
 		4BC821A11C29276C00F68E70 /* RainbowTests.swift in Sources */ = {isa = PBXBuildFile; fileRef = 4BC8219F1C29276C00F68E70 /* RainbowTests.swift */; };
 		D118F24F25EF13B700C7B074 /* ColorTests.swift in Sources */ = {isa = PBXBuildFile; fileRef = D118F24E25EF13B700C7B074 /* ColorTests.swift */; };
+		D13C24C12E20F85B000690E1 /* PerformanceTests.swift in Sources */ = {isa = PBXBuildFile; fileRef = D13C24C02E20F85B000690E1 /* PerformanceTests.swift */; };
+		D13C24C32E20F8A7000690E1 /* StyledStringBuilder.swift in Sources */ = {isa = PBXBuildFile; fileRef = D13C24C22E20F8A7000690E1 /* StyledStringBuilder.swift */; };
+		D13C24C42E20F8A7000690E1 /* StyledStringBuilder.swift in Sources */ = {isa = PBXBuildFile; fileRef = D13C24C22E20F8A7000690E1 /* StyledStringBuilder.swift */; };
+		D13C24C52E20F8A7000690E1 /* StyledStringBuilder.swift in Sources */ = {isa = PBXBuildFile; fileRef = D13C24C22E20F8A7000690E1 /* StyledStringBuilder.swift */; };
+		D13C24C62E20F8A7000690E1 /* StyledStringBuilder.swift in Sources */ = {isa = PBXBuildFile; fileRef = D13C24C22E20F8A7000690E1 /* StyledStringBuilder.swift */; };
+		D13C24CA2E20FC16000690E1 /* StyledStringBuilderTests.swift in Sources */ = {isa = PBXBuildFile; fileRef = D13C24C92E20FC16000690E1 /* StyledStringBuilderTests.swift */; };
 		D16B37C41C2A2F76001A5D35 /* OutputTarget.swift in Sources */ = {isa = PBXBuildFile; fileRef = D16B37C31C2A2F76001A5D35 /* OutputTarget.swift */; };
 		D17B42F025E891D5002914A0 /* ConsoleTextParserTests.swift in Sources */ = {isa = PBXBuildFile; fileRef = D17B42EF25E891D5002914A0 /* ConsoleTextParserTests.swift */; };
 		D1A98FAF1C299D64000DDF20 /* String+Rainbow.swift in Sources */ = {isa = PBXBuildFile; fileRef = D1A98FAE1C299D63000DDF20 /* String+Rainbow.swift */; };
@@ -84,6 +90,9 @@
 		4BC8219E1C29276C00F68E70 /* Info.plist */ = {isa = PBXFileReference; fileEncoding = 4; lastKnownFileType = text.plist.xml; name = Info.plist; path = Tests/Info.plist; sourceTree = SOURCE_ROOT; };
 		4BC8219F1C29276C00F68E70 /* RainbowTests.swift */ = {isa = PBXFileReference; fileEncoding = 4; lastKnownFileType = sourcecode.swift; name = RainbowTests.swift; path = Tests/RainbowTests/RainbowTests.swift; sourceTree = SOURCE_ROOT; };
 		D118F24E25EF13B700C7B074 /* ColorTests.swift */ = {isa = PBXFileReference; fileEncoding = 4; lastKnownFileType = sourcecode.swift; name = ColorTests.swift; path = RainbowTests/ColorTests.swift; sourceTree = "<group>"; };
+		D13C24C02E20F85B000690E1 /* PerformanceTests.swift */ = {isa = PBXFileReference; lastKnownFileType = sourcecode.swift; name = PerformanceTests.swift; path = RainbowTests/PerformanceTests.swift; sourceTree = "<group>"; };
+		D13C24C22E20F8A7000690E1 /* StyledStringBuilder.swift */ = {isa = PBXFileReference; lastKnownFileType = sourcecode.swift; path = StyledStringBuilder.swift; sourceTree = "<group>"; };
+		D13C24C92E20FC16000690E1 /* StyledStringBuilderTests.swift */ = {isa = PBXFileReference; lastKnownFileType = sourcecode.swift; name = StyledStringBuilderTests.swift; path = RainbowTests/StyledStringBuilderTests.swift; sourceTree = "<group>"; };
 		D16B37C31C2A2F76001A5D35 /* OutputTarget.swift */ = {isa = PBXFileReference; fileEncoding = 4; lastKnownFileType = sourcecode.swift; path = OutputTarget.swift; sourceTree = "<group>"; };
 		D17B42EF25E891D5002914A0 /* ConsoleTextParserTests.swift */ = {isa = PBXFileReference; lastKnownFileType = sourcecode.swift; name = ConsoleTextParserTests.swift; path = RainbowTests/ConsoleTextParserTests.swift; sourceTree = "<group>"; };
 		D1A98FAE1C299D63000DDF20 /* String+Rainbow.swift */ = {isa = PBXFileReference; fileEncoding = 4; lastKnownFileType = sourcecode.swift; path = "String+Rainbow.swift"; sourceTree = "<group>"; };
@@ -166,6 +175,8 @@
 				D17B42EF25E891D5002914A0 /* ConsoleTextParserTests.swift */,
 				D1A98FB01C299DF0000DDF20 /* ConsoleStringTests.swift */,
 				D1B428C825F1CBA0009F1DD1 /* ColorApproximatedTests.swift */,
+				D13C24C02E20F85B000690E1 /* PerformanceTests.swift */,
+				D13C24C92E20FC16000690E1 /* StyledStringBuilderTests.swift */,
 			);
 			path = Tests;
 			sourceTree = "<group>";
@@ -181,6 +192,7 @@
 		4BC821A91C2928CF00F68E70 /* Sources */ = {
 			isa = PBXGroup;
 			children = (
+				D13C24C22E20F8A7000690E1 /* StyledStringBuilder.swift */,
 				D1C5D8DB1C2ACA1D001CB619 /* CodesParser.swift */,
 				D1C5D8DD1C2ACA7A001CB619 /* ModesExtractor.swift */,
 				D1C5D8DF1C2AD38A001CB619 /* StringGenerator.swift */,
@@ -386,6 +398,7 @@
 				D1B428C025F1BE56009F1DD1 /* ColorApproximation.swift in Sources */,
 				4B1929191C2BBC98005B79AB /* CodesParser.swift in Sources */,
 				4B19291A1C2BBC98005B79AB /* ModesExtractor.swift in Sources */,
+				D13C24C52E20F8A7000690E1 /* StyledStringBuilder.swift in Sources */,
 				4B19291B1C2BBC98005B79AB /* StringGenerator.swift in Sources */,
 				4B19291C1C2BBC98005B79AB /* Rainbow.swift in Sources */,
 				4B19291D1C2BBC98005B79AB /* Color.swift in Sources */,
@@ -404,6 +417,7 @@
 				D1B428C125F1BE56009F1DD1 /* ColorApproximation.swift in Sources */,
 				4B1929311C2BBE06005B79AB /* CodesParser.swift in Sources */,
 				4B1929321C2BBE06005B79AB /* ModesExtractor.swift in Sources */,
+				D13C24C62E20F8A7000690E1 /* StyledStringBuilder.swift in Sources */,
 				4B1929331C2BBE06005B79AB /* StringGenerator.swift in Sources */,
 				4B1929341C2BBE06005B79AB /* Rainbow.swift in Sources */,
 				4B1929351C2BBE06005B79AB /* Color.swift in Sources */,
@@ -422,6 +436,7 @@
 				D1B428C225F1BE56009F1DD1 /* ColorApproximation.swift in Sources */,
 				4B19295A1C2BBEFE005B79AB /* CodesParser.swift in Sources */,
 				4B19295B1C2BBEFE005B79AB /* ModesExtractor.swift in Sources */,
+				D13C24C42E20F8A7000690E1 /* StyledStringBuilder.swift in Sources */,
 				4B19295C1C2BBEFE005B79AB /* StringGenerator.swift in Sources */,
 				4B19295D1C2BBEFE005B79AB /* Rainbow.swift in Sources */,
 				4B19295E1C2BBEFE005B79AB /* Color.swift in Sources */,
@@ -440,6 +455,7 @@
 				D1B428BF25F1BE56009F1DD1 /* ColorApproximation.swift in Sources */,
 				4BC026EB1C292B25009FD2FD /* ControlCode.swift in Sources */,
 				4BC026ED1C292F5C009FD2FD /* BackgroundColor.swift in Sources */,
+				D13C24C32E20F8A7000690E1 /* StyledStringBuilder.swift in Sources */,
 				4BC026E91C2929E7009FD2FD /* Color.swift in Sources */,
 				D1A98FAF1C299D64000DDF20 /* String+Rainbow.swift in Sources */,
 				D1C5D8E01C2AD38A001CB619 /* StringGenerator.swift in Sources */,
@@ -457,7 +473,9 @@
 			files = (
 				D1B428C925F1CBA0009F1DD1 /* ColorApproximatedTests.swift in Sources */,
 				D17B42F025E891D5002914A0 /* ConsoleTextParserTests.swift in Sources */,
+				D13C24C12E20F85B000690E1 /* PerformanceTests.swift in Sources */,
 				D118F24F25EF13B700C7B074 /* ColorTests.swift in Sources */,
+				D13C24CA2E20FC16000690E1 /* StyledStringBuilderTests.swift in Sources */,
 				D1A98FB11C299DF0000DDF20 /* ConsoleStringTests.swift in Sources */,
 				4BC821A11C29276C00F68E70 /* RainbowTests.swift in Sources */,
 			);