… PERFORMANCE_GUIDE.md
… Tests/RainbowTests/StyledStringBuilderTests.swift